Rename json_api -> json_api_params
